Refactor: util 패키지 분리 #1
